% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiOrderConfirm.R
\name{EdiOrderConfirmUI}
\alias{EdiOrderConfirmUI}
\title{EDI销售订单查询生成器界面}
\usage{
EdiOrderConfirmUI(
  tabTitle = "EDI销售订单确认",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = EdiOrderConfirmUI_left,
  func_right = EdiOrderConfirmUI_right,
  func_bottom = EdiOrderConfirmUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
EDI销售订单查询生成器界面
}
\examples{
EdiOrderConfirmUI()
}
